+++ apis/dbaas/postgres/v1alpha1/zz_generated.managedlist.go
@@ -28,6 +28,15 @@ func (l *PostgresClusterList) GetItems() []resource.Managed {
 	return items
 }
 
+// GetItems of this PostgresDatabaseList.
+func (l *PostgresDatabaseList) GetItems() []resource.Managed {
+	items := make([]resource.Managed, len(l.Items))
+	for i := range l.Items {
+		items[i] = &l.Items[i]
+	}
+	return items
+}
+
 // GetItems of this PostgresUserList.
 func (l *PostgresUserList) GetItems() []resource.Managed {
 	items := make([]resource.Managed, len(l.Items))