+++ docs/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.15] (July 2024)
+- **Features**:
+    - Add `database` CRD for postgres
+    - Upgrade crossplane-runtime to 1.16.0
+    - Shortnames for some of the resources. E.g. `alb` for `applicationloadbalancer`
 ## [1.0.14] (June 2024)
 - **Fixes**:
     - Failing unit tests for compute user